+++ src/globalOptions.ts
@@ -25,10 +25,6 @@ export function setGlobalOptions(options: IGlobalOptions) {
 export function parseENV(): void {
   logger.info('"parseENV" got called');
 
-  if (process.env.TG_USE_NEW_ENUM?.length > 0) {
-    logger.warn('TG_USE_NEW_ENUM & useNewEnum got deprecated, see changelog 6.2 for more');
-  }
-
   const options: IGlobalOptions = {
     globalOptions: {},
     options: {