+++ src/typegoose.ts
@@ -279,7 +279,7 @@ export function getDiscriminatorModelForClass<T, U extends AnyParamConstructor<T
 
   const discriminatorKey = sch.get('discriminatorKey');
   if (sch.path(discriminatorKey)) {
-    sch.paths[discriminatorKey].options.$skipDiscriminatorCheck = true;
+    (sch.paths[discriminatorKey] as any).options.$skipDiscriminatorCheck = true;
   }
 
   const model = from.discriminator(name, sch, id ? id : name);