+++ github-page/changelog.md
@@ -36,6 +36,8 @@ redirect_from:
 - `utils`'s `getName` function now gets exported
 - Remove Deprecated value `overwrite` for `VirtualOptions`
 - Remove instance properties from Model type (remove `& T` from `ModelType`)
+- [IC] Completly remove `TG_USE_NEW_ENUM` from documentation & code
+- [IC] Replace almost all if-throw with the internal `assertion` function
 - [IC] Move VirtualPopulate cache to Reflection
 - [IC] Move Plugins cache to Reflection
 - [IC] Move Hooks cache to Reflection